Move auth methods to separate class
